fix auth token packet parse

+++ telemetry/telemetry.go
@@ -124,14 +124,14 @@ var (
 
 // PositionInfo - position info (base check content)
 type PositionInfo struct {
-	Time int64 `json:"time"`
-	Auth *Auth `json:"auth"`
+	Time float64 `json:"time"`
+	Auth *Auth   `json:"auth"`
 }
 
 // Auth - auth token info
 type Auth struct {
-	Token   string `json:"token"`
-	Expired uint64 `json:"expired"`
+	Token   string  `json:"token"`
+	Expired float64 `json:"expired"`
 }
 
 type ParamsFloat32 struct {
